feat: add icon retailer solid
