+++ packages/ocean-icons-react/src/index.ts
@@ -180,6 +180,7 @@ export { default as ReceiptTax } from './ReceiptTax';
 export { default as Refer } from './Refer';
 export { default as Refresh } from './Refresh';
 export { default as Reply } from './Reply';
+export { default as Retailer } from './Retailer';
 export { default as Rewind } from './Rewind';
 export { default as Rss } from './Rss';
 export { default as SaveAs } from './SaveAs';
@@ -431,7 +432,6 @@ export { default as ReceiptTaxOutline } from './ReceiptTaxOutline';
 export { default as ReferOutline } from './ReferOutline';
 export { default as RefreshOutline } from './RefreshOutline';
 export { default as ReplyOutline } from './ReplyOutline';
-export { default as RetailerOutline } from './RetailerOutline';
 export { default as RewindOutline } from './RewindOutline';
 export { default as RssOutline } from './RssOutline';
 export { default as SaveAsOutline } from './SaveAsOutline';
